mv converter scripts to conv.ts
